Fix: changed header structure added tooltip for inactive nav-menu
